feature: list with search function
