+++ frontend/src/App.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useState } from "react";
-import CurrencyList from "./components/CurrencyList";
+import CurrencyDropdown from "./components/CurrencyDropdown.tsx";
+import CurrencyList from "./components/CurrencyList.tsx";
 import type { Currencies, Currency } from "./types.tsx";
 
 function App() {
@@ -120,7 +121,7 @@ function App() {
               }}
             />
 
-            <CurrencyList
+            <CurrencyDropdown
               value={fromCurrency.isoCode}
               currencies={currencies}
               onChange={setFromCurrency}
@@ -160,7 +161,7 @@ function App() {
               }
             />
 
-            <CurrencyList
+            <CurrencyDropdown
               value={toCurrency.isoCode}
               currencies={currencies}
               onChange={setToCurrency}
@@ -186,42 +187,8 @@ function App() {
         </form>
       </div>
 
-      {/*Currencies*/}
-      <div className="flex justify-center">
-        <div
-          className={`w-full sm:w-[90%] md:w-[60%] lg:w-[35%] mt-8 pt-4 pb-8 px-8 rounded-3xl shadow-lg
-                    ${darkMode ? "bg-[#2A2544]" : "bg-[#F5F3FF]"}`}
-        >
-          <label
-            className={`py-2 block text-sm ${darkMode ? "text-[#C4B5FD]" : "text-[#4C1D95]"}`}
-          >
-            Currency name or ISO code
-          </label>
-          <input
-            type="text"
-            className={`mb-6 rounded-md w-full pl-2 py-2 outline-none
-                            ${
-                              darkMode
-                                ? "bg-[#1E1B2E] text-white focus:ring-[#7C3AED]"
-                                : "bg-white focus:ring-[#C4B5FD]"
-                            }`}
-            placeholder="Currency name or ISO code..."
-            onChange={() => {}}
-          />
-          <ul>
-            {currencies.map((c) => (
-              <li
-                key={c.isoCode}
-                className={`flex justify-between border-b-2 py-2
-                ${darkMode ? "border-[#7C3AED] text-[#DDD6FE]" : "border-[#C4B5FD]"}`}
-              >
-                <p>{c.isoCode}</p>
-                <p>{c.name}</p>
-              </li>
-            ))}
-          </ul>
-        </div>
-      </div>
+      {/*CurrencyList*/}
+      <CurrencyList currencies={currencies} darkMode={darkMode} />
     </div>
   );
 }
